Juno gui()
It should be more clear that `gui()` still exists in Juno, but it's not the default behavior. I think it read before that the only way to get the standalone plot window (when in Juno) would be to change the environment variable. I know it's seems redundant with what's above, but I think this will stop all confusion.