Document, to not use return in recipes
Currently, using the return keyword in a recipe function body is not supported, see https://github.com/JuliaPlots/RecipesBase.jl/issues/58.
This PR documents this. 
I am not sure, whether the MethodError is the only error which can be generated by the use of return, but this is what I have encountered and the type information is the type of what I try to return. 
Should we add a link to the RecipesBase Issue in the documentation?