Document RecipesBase 0.9 requirement
I converted the mention on top into a compat box, to clarify, that using the return keyword requires RecipesBase v0.9. I am not sure,about the version number. Feel free to change this, if this will be tagged as v0.8.1. 
I suggest to keep also the Error Message and the tip section, because it doesn't hurt and might help people which are stuck on an older RecipesBase release, or just don't realize this as the problem.